This is the commit if the UVW notebook
This is a rework of the UVW notebook. Bugs have been fixed in this namely:
The circumpolar bug
The setting time is before the rise time bug
Hard coded values in the plot titles.

Additions include:
The target source is easy to change
The observation duration can be set
Checks to ensure that observation duration is changed if it to long.
There are now instructions.
And the notebook varibles can be set in the cell after the instructions
There has been a fix of the UVW calculation loop that was missing the first and last antennas
This has been optimised so that only half the points are calculated and in the plotting the points are mirrored